add note for troubleshooting poetry installs
Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

Poetry seems to require using the system keyring when installing
dependencies. This can cause the install to hang if for whatever
reason Poetry isn't able to access the keyring.

This PR adds a brief note to the README with a workaround for that
case, setting the `POETRY_KEYRING_ENABLED` envvar to False.